// client-id/src/app/core/models/inspection-popup/ipt-heatmeter-model.ts
export class IptHeatMeterModel{
    athrzSn: number;
    athrzDetailSn: number;
    fluxFomCode: string;
    fluxFomCodeName: string;
    innerFomCode: string;
    innerFomCodeName: string;
    kndCode: string;
    kndCodeName: string;
    fluxLevelCode: string;
    fluxLevelCodeName: string;
    fluxTpLimitNm: string;
    fluxNmNmValue: number;
    fluxEnvrnGradCode: string;
    fluxEnvrnGradCodeName: string;
    calcTpLimitNm: string;
    calcEnvrnGradCode: string;
    calcEnvrnGradCodeName: string;
    calcTmstFomNm: string;
    calcMainSuplySrcelctValue: number;
    calcAsstnSuplySrcelctValue: number;
    overFluxVal: number;
    mxmmFluxVal: number;
    mummFluxVal: number;
    mode: string;
    constructor(){
        this.athrzSn = 0;
        this.athrzDetailSn = 0;
        this.fluxFomCode = null;
        this.fluxFomCodeName = null;
        this.innerFomCode = null;
        this.innerFomCodeName = null;
        this.kndCode = null;
        this.kndCodeName = null;
        this.fluxLevelCode = null;
        this.fluxLevelCodeName = null;
        this.fluxTpLimitNm = null;
        this.fluxNmNmValue = 0;
        this.fluxEnvrnGradCode = null;
        this.fluxEnvrnGradCodeName = null;
        this.calcTpLimitNm = null;
        this.calcEnvrnGradCode = null;
        this.calcEnvrnGradCodeName = null;
        this.calcTmstFomNm = null;
        this.calcMainSuplySrcelctValue = 0;
        this.calcAsstnSuplySrcelctValue = 0;
        this.overFluxVal = 0;
        this.mxmmFluxVal = 0;
        this.mummFluxVal = 0;
        this.mode = null;
    }
}